refactor(limitedReservation): 예약 기반 주문 도메인을 Order 로 통합
