+++ src/main/java/com/example/ililbooks/domain/limitedreservation/service/LimitedReservationService.java
@@ -7,6 +7,8 @@ import com.example.ililbooks.domain.limitedreservation.dto.response.LimitedReser
 import com.example.ililbooks.domain.limitedreservation.entity.LimitedReservation;
 import com.example.ililbooks.domain.limitedreservation.enums.LimitedReservationStatus;
 import com.example.ililbooks.domain.limitedreservation.repository.LimitedReservationRepository;
+import com.example.ililbooks.domain.order.dto.response.OrdersGetResponse;
+import com.example.ililbooks.domain.order.entity.Order;
 import com.example.ililbooks.domain.user.entity.Users;
 import com.example.ililbooks.domain.user.repository.UserRepository;
 import com.example.ililbooks.global.dto.AuthUser;
@@ -104,10 +106,55 @@ public class LimitedReservationService {
         reservation.markCanceled();
     }
 
+    /*
+     * 성공한 예약 기반 주문 생성
+     */
+    @Transactional
+    public OrdersGetResponse createOrderForReservation(AuthUser authUser, Long reservationId) {
+        LimitedReservation reservation = validateReservation(authUser, reservationId);
+        LimitedEvent limitedEvent = reservation.getLimitedEvent();
+
+        if (reservation.getExpiresAt().isBefore(Instant.now())) {
+            throw new BadRequestException(RESERVATION_EXPIRED.getMessage());
+        }
+
+        if (reservation.hasOrder()) {
+            throw new BadRequestException(ALREADY_ORDERED.getMessage());
+        }
+
+        if (limitedEvent.getBookQuantity() < 1) {
+            throw new BadRequestException(OUT_OF_STOCK.getMessage());
+        }
+
+        limitedEvent.decreaseBookQuantity(1);
+
+        Users user = findUser(authUser.getUserId());
+        Order order = Order.of(user, limitedEvent.getBook().getPrice());
+        reservation.linkOrder(order);
+
+        return OrdersGetResponse.of(order);
+    }
+
     // --- 내부 메서드 ---
     private void validateNotAlreadyReserved(Long userId, LimitedEvent event) {
         limitedReservationRepository.findByUsersIdAndLimitedEvent(userId, event)
-                .ifPresent(r -> { throw new BadRequestException(ALREADY_RESERVED_EVENT.getMessage()); });
+                .ifPresent(r -> {
+                    throw new BadRequestException(ALREADY_RESERVED_EVENT.getMessage());
+                });
+    }
+
+    private LimitedReservation validateReservation(AuthUser authUser, Long reservationId) {
+        LimitedReservation reservation = findReservation(reservationId);
+
+        if (!reservation.getUsers().getId().equals(authUser.getUserId())) {
+            throw new BadRequestException(NO_PERMISSION.getMessage());
+        }
+
+        if (reservation.getStatus() != LimitedReservationStatus.SUCCESS) {
+            throw new BadRequestException(RESERVATION_NOT_SUCCESS.getMessage());
+        }
+
+        return reservation;
     }
 
     private LimitedEvent findEvent(Long limitedEventId) {
